telegram-bot/requirements.txt
bitarray==2.8.5
certifi==2023.11.17
charset-normalizer==3.3.2
cytoolz==0.12.2
dnspython==2.4.2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
hexbytes==0.3.1
idna==3.6
parsimonious==0.9.0
pycryptodome==3.19.0
pymongo==4.6.1
pyTelegramBotAPI==4.14.0
regex==2023.10.3
requests==2.31.0
rlp==4.0.0
telebot==0.0.5
toolz==0.12.0
typing_extensions==4.9.0
urllib3==2.1.0
